In sblj, made the return value for LO setting to be a tuple of both the constant number and the name. This way it can easily be reported.
In ovro, made it so that we can at most take 2 antennas. This was necessary because of the way that the data is to be transferred. The transfer mechanism needs to be strictly defined in terms of size and by locking in to at most 2 antennas, it made it simpler to handle. (If there ends up being more than 2 antennas, refactoring will not be difficult.)